Initialize all buffers on start
Previously, buffers were only initialized when they were opened.
However, after testing with hundreds of files, I can see that there was
no real impact on startup time, and given the small size of the expected
files, I don't think it would have a major impact on memory usage. This
allows for multi-file searches to be done in the future.