Reduce CPU usage even more by switching nb_getch() for getch()
Since the screen only updates on user input, there was really no need to
use a non-blocking function. Instead of idling on about 2.0% CPU usage
(according to htop) with nb_getch() and msleep(), navipage idles at
about 0.0% CPU usage, and jumps only to 0.6% when scrolling with j/k
keys. It still jumps to about 6.0% when scrolling with the mouse, and I
was worried about this, until I tried the same thing in nvim. It jumped
to 46.0% usage! So I am not worried about this. Idle CPU usage was my
main worry anyway.