Fix very serious CPU usage leak upon some exits
This fixes a bug where closing the terminal window where navipage was
running would not actually close navipage, but instead bring CPU usage
to 100%.

The cause of this bug was my misunderstanding of how the signal()
function works. I thought that if a signal was handled with a function
designated with signal() that that funciton would execute, and then if
the signal would have exited the program, then it would do so. It is
clear to me now that this is not the case, and that the program never
exited when terminated this way.

This bug was fixed by instead handling the various termination signals
with quit() instead of cleanup(). quit() still calls cleanup(), but the
program is now explicitly exited with proper error codes.