Define OPTSTRING as a constant in config.mk
This has the advantage that if/when I add new options, I need only
modify this one variable in config.mk. This is an improvement over
having to modify
	- the USAGE string
	- the getopt(3) sequence
	- the man page

This has the potential future disadvantage of being infeasible if I ever
add arguments to options, which would require a string passed to getopt
like "xyo:z" or something like that. This option would then need to be
documented differently in USAGE and in the man page, like
"[-xyz] [-o arg]".

This also doesn't handle the documentation in README.md, but I am going
to be changing that soon anyway.