Make README.md document different things than the man page
Previously, README.md was just a Markdown transcription of the man page.
This was done because I thought that it would be best if the same
information about the project could be easily read by those who have the
package installed and those who don't. However, I now realize the
different purposes that each format can serve.