Remove errsv variables and just use errno
A temporary variable should be used if there is a function call made
between when the function you are printing an error for failed and when
you print the error that could have modified errno. But this was not the
case in any of the places where errsv was used.