Use single angle brackets for links
I had two present so that angle brackets would appear in the web view of
this file. I wanted this because I like to view links with angle
brackets around them in the terminal, and I thought it would be good to
see them in the web view of README.md as well. I now realize this to be
a wrong thing to do, because of the difference in the way links are
presented on the web compared to a file viewed in a temrinal. It is fine
with me that links appear simply as links in the web view of the file.