Remove basename() holder variables
In this context, basename() returns a pointer to a char within
copy[0]/copy[1]. There is no need to store these in a variable, and
indeed it seems like GCC optimizes it out.

Now that I think about it, the presence of these variables is probably a
remnant of the misunderstandings that were fixed in 62a6c631.

The comment explaining ret doesn't make sense without these variables
present anymore, so it is removed too.